Enable ^full low level driver'.
I've just discovered another option! The ^full low level driver' adds more defines and some helper functions. It appears to be half way between the ^low level driver^ and the HAL.

It appears some of the defines in the template 'stm32l4xx.h' that I removed because I didn't know what they were for, are related to the ^full low level driver^.