Correctly enable NSS control via the SPI peripheral.
The SPI peripheral appears to expect the NSS output to be open-drain with a pull-up. The internal pull-ups are pretty weak but I can see the correct thing happening.

'NSS pulse management' is a special mode that we don't want. With this disabled I believe NSS is asserted when the SPI peripheral is enabled and de-asserted when SPI is disabled.