Explicitly define a section for the 'user heap and stack'.
I copied this from the example linker script.

Initially I thought it was necessary because the comments seemed to imply it was just a check. When a I tried to use 'printf' I discovered that the 'end' symbol is referenced in libnosys.a. I didn't notice before because I was trying to link the ^correct^ functions.

Possibly the heap and stack check symbols are really needed but they seem like a reasonble thing to have.