Put build output files in a sub-directory.
Keep source directory tidy by putting build output files in a sub-directory.

But, also, I'm thinking of creating different build configurations and thought it would help if the different configurtions used different output directories.

The linker script was naming the input files explicitly. This broke when the object file was in a sub-directory because the name includes the sub-directory. Using a wild card seems normal and is more flexible.